Create DataValue, OrgUnit domain entities
