+++ src/domain/entities/DataValue.ts
@@ -1,0 +1,14 @@
+import { Id } from "./ReferenceObject";
+
+export interface DataValue {
+    dataElement: Id;
+    value: string;
+    orgUnit: Id;
+    period: string;
+    attributeOptionCombo?: Id;
+    categoryOptionCombo?: Id;
+}
+
+export interface DataValueSet {
+    dataValues: DataValue[];
+}
