+++ src/domain/entities/OrgUnit.ts
@@ -1,0 +1,7 @@
+import { Id } from "./ReferenceObject";
+
+export interface OrgUnit {
+    id: Id;
+    featureType?: "NONE" | "MULTI_POLYGON" | "POLYGON" | "POINT" | "SYMBOL";
+    coordinates?: string;
+}
